refactor: extract themeLoader func
